+++ familyhelper-life-sdk/src/main/java/com/dwarfeng/familyhelper/life/sdk/util/Constants.java
@@ -47,6 +47,11 @@ public final class Constants {
     @RemindScopeTypeItem
     public static final int REMIND_SCOPE_TYPE_ALL_PERMITTED = 2;
 
+    /**
+     * @since 1.1.1
+     */
+    public static final int IO_TRANS_BUFFER_SIZE = 4096;
+
     private static final Logger LOGGER = LoggerFactory.getLogger(Constants.class);
     private static final Lock LOCK = new ReentrantLock();